+++ juriscraper/state/texas/common.py
@@ -24,6 +24,19 @@ from juriscraper.lib.string_utils import (
 )
 
 
+class CourtType(Enum):
+    """Standardized IDs for Texas court types."""
+
+    SUPREME = "texas_final"
+    APPELLATE = "texas_appellate"
+    DISTRICT = "texas_district"
+    PROBATE = "texas_probate"
+    BUSINESS = "texas_business"
+    COUNTY = "texas_county"
+    MUNICIPAL = "texas_municipal"
+    UNKNOWN = "texas_unknown"
+
+
 class CourtID(Enum):
     """
     Standardized IDs for Texas courts. Created by prefixing "texas_" to the ID
@@ -182,18 +195,22 @@ class TexasCaseParty(TypedDict):
     representatives: list[str]
 
 
-class TexasTrialCourt(TypedDict):
+class TexasOriginatingCourt(TypedDict):
     """
-    Schema for Texas Trial Court details.
+    Schema for Texas Originating Court details.
 
-    This class is a `TypedDict` that defines the schema for representing a
-    Texas Trial Court. It is used as a utility for safety and type hints.
+    This class is a `TypedDict` that defines the schema for representing the
+    original court to hear a case. This is typically a trial court, but in some
+    cases is an appellate or even last resort court.
 
     The `judge`, `reporter`, and `punishment` attributes may be empty (most
     often the `punishment` field will be empty in civil cases), and it is up to
     the consumer to handle those cases.
 
     :ivar name: The name of the court.
+    :ivar court_id: The ID of the court. Only populated for appellate and
+    last resort courts.
+    :ivar court_type: The type of the court.
     :ivar county: The county where the court is located.
     :ivar judge: The name of the presiding judge in the court.
     :ivar case: Specific case identification or name handled by this court.
@@ -203,12 +220,36 @@ class TexasTrialCourt(TypedDict):
     """
 
     name: str
-    court_id: str
+    court_type: str
     county: str
     judge: str
     case: str
     reporter: str
     punishment: str
+
+
+class TexasOriginatingAppellateCourt(TexasOriginatingCourt):
+    """
+    Schema for Texas Originating Court details when that court is an appellate
+    court.
+
+    Includes the court ID for convenience.
+
+    :ivar court_id: The ID of the appellate court.
+    """
+
+    court_id: str
+
+
+class TexasOriginatingDistrictCourt(TexasOriginatingCourt):
+    """
+    Schema for Texas Originating Court details when that court is a district
+    court.
+
+    :ivar district: The district where the court is located.
+    """
+
+    district: int
 
 
 DISTRICT_COURT_RE = re.compile(
@@ -219,7 +260,7 @@ DISTRICT_COURT_DISTRICT_RE = re.compile(r"^(\d+)\w*$")
 BUSINESS_COURT_RE = re.compile(r"^Business\s+Court", re.IGNORECASE)
 
 
-def _trial_court_name_to_id(name: str, county: str) -> str:
+def _originating_court_name_to_type(name: str) -> CourtType:
     """Takes in a trial court name from TAMES and returns the ID in
     Courtlistener.
 
@@ -228,40 +269,43 @@ def _trial_court_name_to_id(name: str, county: str) -> str:
     Court" for a case can be a Court of Appeals.
 
     :param name: The trial court name from TAMES.
-    :param county: The county where the court is located.
     :return: The CL ID of the trial court."""
     name = name.lower().strip()
+    if not name:
+        return CourtType.UNKNOWN
     # District courts
     district_court_match = DISTRICT_COURT_RE.match(name)
     if district_court_match is not None:
-        district = district_court_match.group(1)
-        if district == "1-a" or district == "1a":
-            return "texdistct2"
-        n = int(DISTRICT_COURT_DISTRICT_RE.match(district).group(1))
-        if n > 1:
-            n += 1
-        return f"texdistct{n}"
+        return CourtType.DISTRICT
 
     # Business court
     business_court_match = BUSINESS_COURT_RE.match(name)
     if business_court_match is not None:
-        return "texbizct"
+        return CourtType.BUSINESS
 
     # Probate courts
     if name.find("probate") >= 0:
-        return "texprobct"
+        return CourtType.PROBATE
 
     # Unknown
     if name.find("unknown court") >= 0:
-        return "unknown"
+        return CourtType.UNKNOWN
 
     # County courts
-    county_slug = county.lower().replace(" ", "_")
     if name.find("municipal") < 0:
-        return f"texcntyct{county_slug}"
+        return CourtType.COUNTY
 
     # Assume anything left is a municipal court
-    return "texctyct"
+    return CourtType.MUNICIPAL
+
+
+def district_court_number_from_name(name: str) -> int:
+    district_court_match = DISTRICT_COURT_RE.match(name)
+    district = district_court_match.group(1)
+    if district == "1-a" or district == "1a":
+        return 1
+    else:
+        return int(DISTRICT_COURT_DISTRICT_RE.match(district).group(1))
 
 
 class TexasCaseDocument(TypedDict):
@@ -330,6 +374,9 @@ class TexasCommonData(TypedDict):
     hints.
 
     :ivar court_id: The ID of the court this docket is from.
+    :ivar court_type: Indicates the type of court this docket is from. The
+    Supreme Court and Court of Criminal Appeals both have unique data and this
+    will indicate if that data is present.
     :ivar docket_number: The docket number of the case.
     :ivar case_name: The shortened and normalized name of the case.
     :ivar case_name_full: The full name of the case.
@@ -337,21 +384,21 @@ class TexasCommonData(TypedDict):
     :ivar case_type: The type of case.
     :ivar parties: A list of parties involved in the case and their associated
     representatives.
-    :ivar trial_court: Information about the trial court handling the case was
-    appealed from.
+    :ivar originating_court: Information about the court a case originated from.
     :ivar case_events: A list of case events (e.g., filing of the case, notice
     of appeal received).
     :ivar appellate_briefs: A list of briefs filed in the case.
     """
 
     court_id: str
+    court_type: str
     docket_number: str
     case_name: str
     case_name_full: str
     date_filed: date
     case_type: str
     parties: list[TexasCaseParty]
-    trial_court: TexasTrialCourt
+    originating_court: TexasOriginatingCourt
     case_events: list[TexasCaseEvent]
     appellate_briefs: list[TexasAppellateBrief]
 
@@ -451,11 +498,12 @@ class TexasCommonScraper(AbstractParser[TexasCommonData]):
 
         data = TexasCommonData(
             court_id=CourtID.UNKNOWN.value,
+            court_type=CourtType.UNKNOWN.value,
             docket_number=self.docket_number,
             date_filed=self._parse_date_filed(),
             case_type=self._parse_case_type(),
             parties=self.parties,
-            trial_court=self._parse_trial_court(),
+            originating_court=self._parse_originating_court(),
             case_events=self._parse_case_events(),
             appellate_briefs=self._parse_appellate_briefs(),
             case_name=self.case_name,
@@ -746,7 +794,13 @@ class TexasCommonScraper(AbstractParser[TexasCommonData]):
             for i in range(n_parties)
         ]
 
-    def _parse_trial_court(self) -> TexasTrialCourt:
+    def _parse_originating_court(
+        self,
+    ) -> (
+        TexasOriginatingCourt
+        | TexasOriginatingAppellateCourt
+        | TexasOriginatingDistrictCourt
+    ):
         """
         Extracts the trial court info from the HTML tree. Will fail if
         `_parse_text` has not yet been called.
@@ -765,16 +819,33 @@ class TexasCommonScraper(AbstractParser[TexasCommonData]):
 
         court_name = fields.get("Court", "")
         county = fields.get("County", "")
+        court_type = _originating_court_name_to_type(court_name)
 
-        return TexasTrialCourt(
+        court_details = TexasOriginatingCourt(
             name=court_name,
-            court_id=_trial_court_name_to_id(court_name, county),
+            court_type=court_type.value,
             county=county,
             judge=fields.get("Court Judge", ""),
             case=fields.get("Court Case", ""),
             reporter=fields.get("Reporter", ""),
             punishment=fields.get("Punishment", ""),
         )
+
+        if court_type == CourtType.APPELLATE:
+            court_id = coa_name_to_court_id(court_name).value
+            originating_court_details = TexasOriginatingAppellateCourt(
+                **court_details,
+                court_id=court_id,
+            )
+        elif court_type == CourtType.DISTRICT:
+            district = district_court_number_from_name(court_name)
+            originating_court_details = TexasOriginatingDistrictCourt(
+                **court_details, district=district
+            )
+        else:
+            originating_court_details = court_details
+
+        return originating_court_details
 
     @staticmethod
     def _parse_case_documents(cell: HtmlElement) -> list[TexasCaseDocument]: